requirements.txt
alembic==1.17.2
aniso8601==10.0.1
Authlib==1.6.5
blinker==1.9.0
cachelib==0.13.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudinary==1.44.1
cryptography==46.0.3
dotenv==0.9.9
Flask==3.1.2
flask-cors==6.0.1
Flask-Email==1.4.4
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-OAuthlib==0.9.6
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
google-auth==2.43.0
greenlet==3.2.4
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
oauthlib==2.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.2
python-dotenv==1.2.1
python-env==1.0.0
pytz==2025.2
redis==7.1.0
requests==2.32.5
requests-oauthlib==1.1.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
